Collision handling
Rigidbody added to capsule (without a rigidbody the OnCollisionEnter function doesn't work and we need that to handle actions that happen when colliding with walls), movement now based on built-in Unity physics (this prevents the rigidbody from phasing through walls).
Collision is now registered as a simple sound in the CollisionHandler script.